+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K9Y89TFVJVCW9JFCQS6518CJ"
+        "01K9YEE44ABM89H41W0B7VKTTQ"
     )
     r"""The unique identifier of the data source"""