+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -121,5 +121,5 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-13T09:21:05.843Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-13T11:08:17.262Z")
     r"""The date and time the resource was last updated"""