+++ packages/orq-rc/src/orq_ai_sdk/tools.py
@@ -3,10 +3,6 @@
 from .basesdk import BaseSDK
 from orq_ai_sdk import models, utils
 from orq_ai_sdk._hooks import HookContext
-from orq_ai_sdk.models import (
-    createtoolop as models_createtoolop,
-    updatetoolop as models_updatetoolop,
-)
 from orq_ai_sdk.types import BaseModel, OptionalNullable, UNSET
 from orq_ai_sdk.utils import get_security_from_env
 from orq_ai_sdk.utils.unmarshal_json_response import unmarshal_json_response
@@ -214,10 +210,7 @@ class Tools(BaseSDK):
         self,
         *,
         request: Optional[
-            Union[
-                models_createtoolop.CreateToolRequestBody,
-                models_createtoolop.CreateToolRequestBodyTypedDict,
-            ]
+            Union[models.CreateToolRequestBody, models.CreateToolRequestBodyTypedDict]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -310,10 +303,7 @@ class Tools(BaseSDK):
         self,
         *,
         request: Optional[
-            Union[
-                models_createtoolop.CreateToolRequestBody,
-                models_createtoolop.CreateToolRequestBodyTypedDict,
-            ]
+            Union[models.CreateToolRequestBody, models.CreateToolRequestBodyTypedDict]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -407,10 +397,7 @@ class Tools(BaseSDK):
         *,
         tool_id: str,
         request_body: Optional[
-            Union[
-                models_updatetoolop.UpdateToolRequestBody,
-                models_updatetoolop.UpdateToolRequestBodyTypedDict,
-            ]
+            Union[models.UpdateToolRequestBody, models.UpdateToolRequestBodyTypedDict]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -518,10 +505,7 @@ class Tools(BaseSDK):
         *,
         tool_id: str,
         request_body: Optional[
-            Union[
-                models_updatetoolop.UpdateToolRequestBody,
-                models_updatetoolop.UpdateToolRequestBodyTypedDict,
-            ]
+            Union[models.UpdateToolRequestBody, models.UpdateToolRequestBodyTypedDict]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,